Maze Levels Update
Decoration, lights and final detail of the levels:

-Down

-Left

-right

-Left-Down

- Right-Down

- Right-Left-Down

(It would be necessary to decide if the enemies applied in said levels are correct, and the drop of items).

- New additions:

A colorful and unique decal has been added for each dropped item for the player to easily locate the item. The problem is that with equipped weapons the decal is visible when the player approaches a wall with the weapon equipped. It would be necessary to repair it or separate the weapon from the ground with a hand weapon.

New props added:

Blinds.
Cables.
Fake doors.

Errors found in the test:
- When enemies die, they create a fairly large collision on the temporary ground for many seconds, preventing the player from walking in that direction.