Update Rooms and Tutorial.
The room is finished:
-Top-Left

Fixed player spawns for all finished rooms so far. Enemies spawn and items have also been added.

Fixed tutorial player critical path collisions.

The neo pixel weapon has been implemented in the tutorial when you turn on the button in the technical room and added medkits and grenades.

The enemies that should be corresponding to the tutorial have been added.

The decal has been changed for our faces